requirements_local.txt
python-dotenv
scrapy
scrapy_zyte_smartproxy

spidermon

w3lib
